Checkin.
I never write good checkin notes here in Xcode when writing iApps. It's probably because so much of it is experimentation, that I get to a point to check something in, and I've done a hundred things, many of which failed.

Here's some stuff. Basic fetching and selectors seem to work. WebKit headless does not work, even in the UI.